Split up pytransform3d.visualizer
